==> analysis/03_TrajectoryAnalysis_scRNA.Rmd <==
---
title: "03_TrajectoryAnalysis"
author: "Afsheen"
date: "2023-01-12"
html_document:
    theme: cerulean
    toc: yes
    toc_float:
      collapsed: true
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

```{r Calling libraries, echo=FALSE, include=FALSE}
library(SingleR)
library(SingleCellExperiment)
library(Seurat)
library(slingshot)
library(ggplot2)
library(tidyverse)
library(viridis)
library(velociraptor)

dir.create("./output/03_Trajectories/", showWarnings = F)
```

```{r}
seurat_integrated <-readRDS("./data/QC_dataClustered.filtTyped.rds")
DefaultAssay(seurat_integrated) <- "RNA"
seurat_integrated@meta.data <- seurat_integrated@meta.data[,!grepl("Lineage|pseudo",colnames(seurat_integrated@meta.data))] 
scRNA.sce <- as.SingleCellExperiment(seurat_integrated)
```

## Calculating Pseudotimes and lineages

```{r Trajectory analysis Calc Pseudtime,echo=FALSE}
sce_sling <- slingshot(scRNA.sce, reducedDim="HARMONY",
                       clusterLabels = scRNA.sce$seurat_clusters, 
                       start.clus=c("3", "9"),
                       end.clus=c("13","11", "2", "6"),
                       omega=T)
```

```{r 08-Trajectory analysis Calc Lineages,echo=FALSE, fig.width=8, fig.height=5}
pseudo.paths <- slingPseudotime(sce_sling)
shared.pseudo <- rowMeans(pseudo.paths, na.rm=TRUE)

embedded <- embedCurves(sce_sling, "UMAP")
embedded <- slingCurves(embedded)

PS<-cbind(colData(scRNA.sce),shared.pseudo, pseudo.paths) %>% as.data.frame()
seurat_integrated@meta.data <- PS
linages= grep("Lineage", colnames(PS), value = T)
names(embedded) = linages
```

## Pseudotime plots by lineage 

```{r,fig.cap="pseudotime UMAP plots", fig.width=12, fig.height=4}
gga <- FeaturePlot(seurat_integrated, "shared.pseudo")+scale_color_viridis()
for (path in embedded) {
  embedded_tmp <- data.frame(path$s[path$ord,])
  gga <- gga + geom_path(data=embedded_tmp, aes(x=umap_1, y=umap_2), linewidth=1.2)
}
gga <- gga + ggtitle("Trajectory-Pseudotime")


ggb <- DimPlot(seurat_integrated, group.by ="seurat_clusters")
for (path in embedded) {
  embedded_tmp <- data.frame(path$s[path$ord,])
  ggb <- ggb + geom_path(data=embedded_tmp, aes(x=umap_1, y=umap_2), linewidth=1.2)
}
ggb <- ggb + ggtitle("Trajectory-Clusters")


# 
ggc <- DimPlot(seurat_integrated, group.by ="Knobl.labels")
for (path in embedded) {
  embedded_tmp <- data.frame(path$s[path$ord,])
  ggc <- ggc + geom_path(data=embedded_tmp, aes(x=umap_1, y=umap_2), linewidth=1.2)
}
ggc <- ggc + ggtitle("Trajectory-KnoblCelltype")

gga|ggb|ggc

ggsave("./output/03_Trajectories/Trajectories_All.svg",gga|ggb|ggc )

```

## Plots for Trajectory Lineages individual

```{r,fig.cap="Trajectory Lineages individual",echo=FALSE, fig.width=12, fig.height=8}
plots=list()
for(lin in linages){
  embedded_tmp = data.frame(embedded[[lin]]$s[embedded[[lin]]$ord,])
  gga <- FeaturePlot(seurat_integrated, features = lin)+scale_color_viridis()+ 
    geom_path(data=embedded_tmp, aes(x=umap_1, y=umap_2), linewidth=1.2)
  plots[[lin]] = gga + ggtitle(lin)
}

p <-ggpubr::ggarrange(plotlist=plots)
p
ggsave("./output/03_Trajectories/Trajectories_Lineages.svg",p)
```



```{r}
saveRDS(file="./data/QC_dataClustered.filtTypedTraj.rds", seurat_integrated)
```
