==> analysis/07_DeconVol_Bulk.Rmd <==
---
title: "Deconvolution"
author: "A Chiocchetti"
date: "2024-01-11"
output:
  html_document:
    theme: cerulean
    toc: yes
    toc_float:
      collapsed: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

```{r}
library(SingleR)
library(DropletUtils)
library(SingleCellExperiment)
library(Seurat)
library(slingshot)
library(gprofiler2)
library(tidyverse)
library(MuSiC)
library(ggstream)
```

```{r}
options(stringsAsFactors = F)
set.seed(157434538)
reanalyze=F
limit_to_markers=F
cutoff = 0.05 # pvalue cutoff for DEG
padj_cu=FALSE # use adjusted pvalue 
Dark8 = RColorBrewer::brewer.pal(8, "Dark2")
```

```{r getdata}
output= paste0("output/07_Bulk_Deconv/")
dir.create(output, showWarnings = F)
source("code/custom_functions.R")
ddsMat <- readRDS("data/Bulk_dds_matrix.rds")

ddsMat <- ddsMat[,ddsMat$PAmM==0]

seurat_integrated <- readRDS("./data/QC_dataClustered.filtTypedTraj.rds")



SampleInfo = as.data.frame(colData(ddsMat))

log_2cpm=log2(counts(ddsMat, normalize=T)+1)
SamplePalette =  c("#000000","#3399FF")

if(limit_to_markers==TRUE){
  Idents(seurat_integrated) <- "seurat_clusters" 
  Markers<-FindAllMarkers(seurat_integrated)
  Mfeatures <- Markers$gene[Markers$p_val_adj<0.05] %>% unique()
}

sce.RNA <- as.SingleCellExperiment(seurat_integrated)
metadata<- seurat_integrated@meta.data

```

```{r}
bulk.mtx = counts(ddsMat)
bulk.meta = colData(ddsMat)
bulk.gene = rowData(ddsMat)
rownames(bulk.mtx) = bulk.gene$hgnc[match(rownames(bulk.mtx), bulk.gene$entrez_gene)]
genes = intersect(rownames(sce.RNA), bulk.gene$hgnc)

if(limit_to_markers==TRUE){
  genes = intersect(genes, Mfeatures)
}

bulk.mtx=bulk.mtx[genes, ]
sce.RNA <- sce.RNA[genes,]

Est.prop.bulk = music_prop(bulk.mtx = bulk.mtx, sc.sce = sce.RNA, 
                           clusters = 'seurat_clusters',
                           samples = 'orig.ident', verbose = F)



prop.matrix = t(data.matrix(Est.prop.bulk$Est.prop.weighted))
```


# proportion organoid clusters (our experiment)

## longitudinal trajectories of proportions

```{r, fig.width=8, fig.height=8}
long_prop_matrix <-reshape2::melt(prop.matrix)
colnames(long_prop_matrix) <-c("Cluster", "Sample", "Proportion")
metadata_bulk <- as.data.frame(colData(ddsMat))
long_prop_matrix <- cbind(long_prop_matrix,metadata_bulk[long_prop_matrix$Sample,])

p1<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion))+
  geom_smooth(method = "loess", alpha = 0.2, span=0.5)+facet_wrap(~Cluster, scales = "free_y")+theme_classic()+scale_x_continuous(breaks=c(5,10,17,40,80))
p1

p2<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion, lty=Batch))+
  geom_smooth(method = "loess", alpha = 0.2, span=0.5)+facet_wrap(~Cluster, scales = "free_y")+theme_classic()+scale_x_continuous(breaks=c(5,10,17,40,80))
p2

p3<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion))+
  geom_smooth(method = "loess", alpha = 0.2, span=0.5)+facet_wrap(~Cluster+Batch, scales = "free_y")+theme_classic()+scale_x_continuous(breaks=c(5,10,17,40,80))
p3


ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters.svg"), p1)
ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters_batchinline.svg"), p2)
ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters_batchcompare.svg"), p3)

```

## testing time differences per timepoint vs timpoint -1


```{r, fig.height=6, fig.width=18}
# To do Cell type proportion per time dev point 

library(lme4)
library(lmerTest)

clustervar = "Cluster"
lmatrix = long_prop_matrix

test_Ratio_by_Day <- function(clustervar, lmatrix){
  res=list()
  Days=sort(unique(SampleInfo$Day))
  for(i in 1:length(Days)){
    if(i==1){} else {
      Dayfilter = SampleInfo$Day %in% Days[c(i-1,i)] & SampleInfo$PAmM==0
      Set = rownames(SampleInfo)[Dayfilter]
      lab = paste("TPrestab", "Day", paste0(Days[c(i-1,i)], collapse = "vs"),"0mM_PA",sep="_")
      print(lab)
      Dataset = lmatrix[lmatrix$Sample %in% Set,]
      Dataset$Day = relevel(as.factor(Dataset$Day), ref = as.character(Days[i-1]))
      for(cl in sort(unique(Dataset[[clustervar]]))){
        ClData <- Dataset[Dataset[[clustervar]]==cl,]
        if(var(ClData$Proportion)==0){
          res_tmp=c(beta=NA, pval=NA)
        }else{
          res_lm<- lmer(Proportion~Day+(1|Batch), data=ClData)
          res_tmp=c(beta=summary(res_lm)$coef[2,"Estimate"], 
                    pval=summary(res_lm)$coef[2,"Pr(>|t|)"])
        }
        res[[paste0(cl,"by",Days[i])]]<- res_tmp
      }
    }
  }
  res<-do.call(rbind, res) %>%  as.data.frame()
  res$cluster <- gsub("by[0-9].","",rownames(res))
  res$Day <- gsub(".*[0-9]*by", "", rownames(res)) %>%  as.numeric()
  res$"sig"<- res$pval<0.05
  res$Day <- as.factor(res$Day)
  return(res)
}


res<- test_Ratio_by_Day(clustervar="Cluster", lmatrix=long_prop_matrix)
res$cluster<- factor(res$cluster, levels=res$cluster[order(res$cluster %>% as.numeric())] %>% unique())

p<- ggplot(res, aes(x=Day, y=beta, fill=sig))+
  geom_bar(stat='identity')+facet_wrap(~cluster, ncol=7, nrow=2)+theme_classic()

p
ggsave(paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_clusters.svg"), p)
openxlsx2::write_xlsx(res, paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_clusters.xlsx"))

```

## Heatmap proportions across sampels and clusters

Columns add up to 1 i.e. 100% of cells 

```{r, fig.height=8, fig.width=6}

col=viridis::viridis(100)

svg(paste0(output, "07_DeconVol_Bulk_propHeatmap_clusters.svg"))
gplots::heatmap.2(prop.matrix, margins = c(15,10), main = "organoid clusters",cexCol = 0.3,
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", key.par = list(mar=c(5,2,5,2)))
dev.off()

gplots::heatmap.2(prop.matrix, margins = c(15,10), main = "organoid clusters", cexCol = 0.3,
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", 
                  key.par = list(mar=c(5,2,5,2)))

```
# proportion knoblich cell types 

```{r}
sceRNAknoblich <- readRDS("data/Knoblich/CHOOSE_ASD_Modules.rds")
Idents(sceRNAknoblich) <- "celltype_cl_coarse"
sceRNAknoblich <- subset(sceRNAknoblich, cells = WhichCells(sceRNAknoblich, downsample = 500))
sceRNAknoblich<- as.SingleCellExperiment(sceRNAknoblich)

Est.prop.bulk.knblich = music_prop(bulk.mtx = bulk.mtx, sc.sce = sceRNAknoblich, 
                                   clusters = 'celltype_cl_coarse',
                                   samples = 'orig.ident', verbose = F)

prop.matrix.knblich = t(data.matrix(Est.prop.bulk.knblich$Est.prop.weighted))

```

## longitudinal trajectories of proportions

```{r}
long_prop_matrix <-reshape2::melt(prop.matrix.knblich)
colnames(long_prop_matrix) <-c("Kn.Cluster", "Sample", "Proportion")
metadata_bulk <- as.data.frame(colData(ddsMat))
long_prop_matrix <- cbind(long_prop_matrix,metadata_bulk[long_prop_matrix$Sample,])

p<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion))+
  geom_smooth(method = "loess", alpha = 0.2,span=0.5 )+facet_wrap(~Kn.Cluster, scales = "free_y")+theme_classic()+scale_x_continuous(breaks=c(5,10,17,40,80))
p
ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_Knoblichclusters.svg"), p)

```
## testing time differences per cell type 

```{r, fig.width=8, fig.height=8}
testres = test_Ratio_by_Day(lmatrix = long_prop_matrix, clustervar = "Kn.Cluster")


p<- ggplot(testres, aes(x=Day, y=beta, fill=sig))+
  geom_bar(stat='identity')+facet_wrap(~cluster)+theme_classic()

p

ggsave(paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_clusters.svg"), p)

openxlsx2::write_xlsx(testres, paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_KnCelltypes.xlsx"))

```

## Heatmap proportions across sampels and clusters

```{r, fig.height=8, fig.width=6}
svg(paste0(output, "07_DeconVol_Bulk_propHeatmap_Knoblichclusters.svg"))
gplots::heatmap.2(prop.matrix.knblich, scale = "col", cexCol = 0.3,
                  margins = c(12,5), main = "knoblich clusters",
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", key.par = list(mar=c(5,2,5,2)))

dev.off()

gplots::heatmap.2(prop.matrix.knblich, scale = "col", cexCol = 0.3, 
                  margins = c(12,5), main = "knoblich clusters",
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", key.par = list(mar=c(5,2,5,2)))

```




# Analyses with merged clusters 

```{r}
DimPlot(seurat_integrated, group.by = "merged_clusters")
```


```{r}

Est.prop.bulk = music_prop(bulk.mtx = bulk.mtx, sc.sce = sce.RNA, 
                           clusters = 'merged_clusters',
                           samples = 'orig.ident', verbose = F)


prop.matrix = t(data.matrix(Est.prop.bulk$Est.prop.weighted))
```


# proportion organoid clusters (our experiment)

## longitudinal trajectories of proportions

```{r, fig.width=8, fig.height=8}
long_prop_matrix <-reshape2::melt(prop.matrix)
colnames(long_prop_matrix) <-c("Cluster", "Sample", "Proportion")
metadata_bulk <- as.data.frame(colData(ddsMat))
long_prop_matrix <- cbind(long_prop_matrix,metadata_bulk[long_prop_matrix$Sample,])

p1<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion))+
  geom_smooth(method = "loess", span=0.5,alpha = 0.2)+facet_wrap(~Cluster, scales = "free_y")+theme_classic()+ scale_x_continuous(breaks=c(5,10,17,40,80))
p1

p2<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion, lty=Batch))+
  geom_smooth(method = "loess", span=0.5,alpha = 0.2)+facet_wrap(~Cluster, scales = "free_y")+theme_classic()+ scale_x_continuous(breaks=c(5,10,17,40,80))
p2

p3<-ggplot(long_prop_matrix, aes(x=Day, y=Proportion))+
  geom_smooth(method = "loess",span=0.5, alpha = 0.2)+facet_wrap(~Cluster+Batch, scales = "free_y")+theme_classic()+ scale_x_continuous(breaks=c(5,10,17,40,80))
p3

ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters_merged.svg"), p1)
ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters_batchinline_merged.svg"), p2)
ggsave(paste0(output, "07_DeconVol_Bulk_timeplots_clusters_batchcompare_merged.svg"), p3)

```

## testing time differences per cluster 

Model: Proportion ~ Day x PamM, corrected for random Batch effects,including Main effects

```{r}


testres = test_Ratio_by_Day(lmatrix = long_prop_matrix, clustervar = "Cluster")


p<- ggplot(testres, aes(x=Day, y=beta, fill=sig))+
  geom_bar(stat='identity')+facet_wrap(~cluster)+theme_classic()

p

ggsave(paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_clusters_merged.svg"), p)

openxlsx2::write_xlsx(testres, paste0(output,"07_DeconVol_Bulk_StatTimecourseDiff_clusters_merged.xlsx"))
```

## Heatmap proportions across sampels and clusters

Columns add up to 1 i.e. 100% of cells 

```{r, fig.height=8, fig.width=6}

col=viridis::viridis(100)

svg(paste0(output, "07_DeconVol_Bulk_propHeatmap_clusters_merged.svg"))
gplots::heatmap.2(prop.matrix, margins = c(15,10), main = "organoid clusters",cexCol = 0.3,
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", key.par = list(mar=c(5,2,5,2)))
dev.off()

gplots::heatmap.2(prop.matrix, margins = c(15,10), main = "organoid clusters", cexCol = 0.3,
                  Colv = F, trace = "none", density.info = "none", col = col, key.title = "prop", key.xlab = "proportion", 
                  key.par = list(mar=c(5,2,5,2)))

```
